backup_20260104/server_requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.11
httpx==0.28.1
